gpu/drm: hisilicon: fix hang issue during power off
When power off display, it will call the dpms callback function for
suspend and power off; in the old code it will call the power down flow
for twice, one time is for suspend and another is for power off; because
the first time has disabled the clock domain for the module, so in the
second time it access the module's register, it will introduce the
external bus error and hang.

So this patch will only turn off the module for power off flow.

Signed-off-by: Leo Yan <leo.yan@linaro.org>